+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 ### Added
 - InvalidQueryException
 - Query Helper traits to assist with resolving user/user group params
+- ActiveRecord service accessor traits
 
 ### Changed
 - Exceptions now inherit Yii's base Exception, not ErrorException